ci_requirements.txt
# catkin_pkg
cflib
num2words
numpy
python-Levenshtein
pyyaml
requests
SpeechRecognition
word2number
# git+https://github.com/davheld/tf#egg=tf
